requirements.txt
google-genai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2